+++ CHANGELOG.md
@@ -123,6 +123,7 @@ In order to resolve various design problems interfering with development, the `a
       In general, to make time pass, create a `Universe` and `step()` it instead.
     - `universe::Handle::new_pending()` has been replaced by `UniverseTransaction::insert()`.
       Handles are now always associated with some container, either `Universe` or `UniverseTransaction`, except for those in the “gone” defunct state.
+    - Behaviors are deprecated. They are no longer supported by `Character` and `Universe`.
     - `math::GridRotation::{CLOCKWISE, COUNTERCLOCKWISE}` have been replaced by `Face6::PY.{clockwise, counterclockwise}()`.
     - `math::Rgb::UNIFORM_LUMINANCE_*` have been replaced by constants on `Rgb01`.
     - `math::Rgb` can no longer be converted to `Block`, `Primitive`, or `Atom`.